+++ practices/cpp/level1/p05_Canvas/Eric/Circle.cpp
@@ -1,0 +1,7 @@
+#include "Circle.h"
+#include <iostream>
+
+void Circle::print() const
+{
+	std::cout << "Circle " << locationX << "," << locationY << " R = " << radius << std::endl;
+}
