+++ practices/cpp/level1/p05_Canvas/Eric/Circle.h
@@ -1,6 +1,5 @@
 #ifndef CIRCLE_H
-#define CIRCEL_H
-#include <iostream>
+#define CIRCLE_H
 #include "Shape.h"
 class Circle : public Shape
 {
@@ -8,12 +7,10 @@ public:
 	Circle() = default;
 	Circle(int sourceX, int sourceY, int sourceR) : 
 		Shape(sourceX, sourceY), radius(sourceR){}
-	void print() const override;
+	virtual void print() const override;
 private:
 	int radius;
 };
-void Circle::print() const
-{
-	std::cout << "Circle " << locationX << "," << locationY << " R = " << radius << std::endl;
-}
+
+
 #endif
