+++ practices/cpp/level1/p05_Canvas/Eric/Rect.cpp
@@ -1,0 +1,5 @@
+#include "Rect.h"
+void Rect::print() const
+{
+	std::cout << "Rect " << locationX << "," << locationY << " length : " << length << " width : " << width << std::endl;
+}
