+++ practices/cpp/level1/p05_Canvas/Eric/Shape.h
@@ -1,6 +1,5 @@
 #ifndef SHAPE_H
 #define SHAPE_H
-
 class Shape
 {
 public:
